tests/wpt/web-platform-tests/workers/Worker_terminate_event_queue.htm

<!DOCTYPE html>
<title> AbstractWorker terminate(): clear event queue </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var testResult;
  var worker = new Worker('./support/WorkerTerminate.js');
  worker.onmessage = function(e) {
    testResult = e.data;
    if (testResult >= 10000) {
      worker.terminate();
      worker.onmessage = t.unreached_func('Unexpected message event');
      setTimeout(t.step_func_done(function() {
        assert_equals(testResult, 10000);
      }), 100);
    }
  };
  worker.postMessage("ping");
});
</script>
